Backend/Soccer3D/requirements.txt
numpy>=1.19.0
opencv-python>=4.5.0
tritonclient[all]>=2.10.0
supervision>=0.6.0
mediapipe>=0.8.9
torch>=1.9.0
torchvision>=0.10.0
pillow>=8.0.0
pyyaml>=5.4.0
matplotlib>=3.3.0
paho-mqtt>=1.6.1
